Address desktop vault page review
